refactor: Create shared type definitions for frontend and backend
- Add shared/types.ts with FeedResult and SearchResult interfaces
- Update worker/types.ts to re-export from shared types
- Update ResultDisplay.tsx to import from shared types
- Add shared directory to tsconfig includes
