+++ tsconfig.app.json
@@ -25,7 +25,9 @@
     "noUnusedParameters": true,
     "erasableSyntaxOnly": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+    "noImplicitReturns": true,
+    "noPropertyAccessFromIndexSignature": true
   },
-  "include": ["src"]
+  "include": ["src", "shared"]
 }
