+++ tsconfig.worker.json
@@ -4,5 +4,5 @@
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.worker.tsbuildinfo",
     "types": ["./worker-configuration.d.ts", "vite/client"]
   },
-  "include": ["./worker-configuration.d.ts", "./worker"]
+  "include": ["./worker-configuration.d.ts", "./worker", "./shared"]
 }
